Add incresing speed, change model ints to doubles

+++ src/main/java/com/codecool/lanpong/models/Ball.java
@@ -2,34 +2,34 @@ package com.codecool.lanpong.models;
 
 public class Ball {
 
-    private int radius;
-    private int xPos;
-    private int yPos;
+    private double radius;
+    private double xPos;
+    private double yPos;
     private double angle;  // 0-360, where 0 is right, 90 is top
 
     public Ball() {
     }
 
-    public Ball(int xPos, int yPos, int radius) {
+    public Ball(double xPos, double yPos, double radius) {
 
         this.xPos = xPos;
         this.yPos = yPos;
         this.radius = radius;
     }
 
-    public int getxPos() {
+    public double getxPos() {
         return xPos;
     }
 
-    public void setxPos(int xPos) {
+    public void setxPos(double xPos) {
         this.xPos = xPos;
     }
 
-    public int getyPos() {
+    public double getyPos() {
         return yPos;
     }
 
-    public void setyPos(int yPos) {
+    public void setyPos(double yPos) {
         this.yPos = yPos;
     }
 
@@ -45,7 +45,7 @@ public class Ball {
         return radius;
     }
 
-    public void setRadius(int radius) {
+    public void setRadius(double radius) {
 
         this.radius = radius;
     }
